+++ core/src/test/java/feign/FeignTest.java
@@ -23,6 +23,8 @@ import okhttp3.mockwebserver.SocketPolicy;
 import okhttp3.mockwebserver.MockWebServer;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import okio.Buffer;
 import org.assertj.core.api.Fail;
@@ -49,6 +51,7 @@ import feign.codec.EncodeException;
 import feign.codec.Encoder;
 import feign.codec.ErrorDecoder;
 import feign.codec.StringDecoder;
+import feign.Feign.ResponseMappingDecoder;
 
 import static feign.Util.UTF_8;
 import static feign.assertj.MockWebServerAssertions.assertThat;
@@ -351,20 +354,36 @@ public class FeignTest {
   }
 
   @Test
-  public void queryMapKeysMustBeStrings() throws Exception {
-    server.enqueue(new MockResponse());
-
+  public void queryMapValueStartingWithBrace() throws Exception {
     TestInterface api = new TestInterfaceBuilder().target("http://localhost:" + server.getPort());
 
-    Map<Object, String> queryMap = new LinkedHashMap<Object, String>();
-    queryMap.put(Integer.valueOf(42), "alice");
+    server.enqueue(new MockResponse());
+    Map<String, Object> queryMap = new LinkedHashMap<String, Object>();
+    queryMap.put("name", "{alice");
+    api.queryMap(queryMap);
+    assertThat(server.takeRequest())
+        .hasPath("/?name=%7Balice");
 
-    try {
-      api.queryMap((Map) queryMap);
-      Fail.failBecauseExceptionWasNotThrown(IllegalStateException.class);
-    } catch (IllegalStateException ex) {
-      assertThat(ex).hasMessage("QueryMap key must be a String: 42");
-    }
+    server.enqueue(new MockResponse());
+    queryMap = new LinkedHashMap<String, Object>();
+    queryMap.put("{name", "alice");
+    api.queryMap(queryMap);
+    assertThat(server.takeRequest())
+        .hasPath("/?%7Bname=alice");
+
+    server.enqueue(new MockResponse());
+    queryMap = new LinkedHashMap<String, Object>();
+    queryMap.put("name", "%7Balice");
+    api.queryMapEncoded(queryMap);
+    assertThat(server.takeRequest())
+        .hasPath("/?name=%7Balice");
+
+    server.enqueue(new MockResponse());
+    queryMap = new LinkedHashMap<String, Object>();
+    queryMap.put("%7Bname", "%7Balice");
+    api.queryMapEncoded(queryMap);
+    assertThat(server.takeRequest())
+        .hasPath("/?%7Bname=%7Balice");
   }
 
   @Test
@@ -565,6 +584,18 @@ public class FeignTest {
   }
 
   @Test
+  public void decodingDoesNotSwallow404ErrorsInDecode404Mode() throws Exception {
+    server.enqueue(new MockResponse().setResponseCode(404));
+    thrown.expect(IllegalArgumentException.class);
+
+    TestInterface api = new TestInterfaceBuilder()
+        .decode404()
+        .errorDecoder(new IllegalArgumentExceptionOn404())
+        .target("http://localhost:" + server.getPort());
+    api.queryMap(Collections.emptyMap());
+  }
+
+  @Test
   public void okIfEncodeRootCauseHasNoMessage() throws Exception {
     server.enqueue(new MockResponse().setBody("success!"));
     thrown.expect(EncodeException.class);
@@ -660,6 +691,49 @@ public class FeignTest {
             .hasPath("/?trim=5.2FSi+");
   }
 
+  @Test
+  public void responseMapperIsAppliedBeforeDelegate() throws IOException {
+    ResponseMappingDecoder decoder = new ResponseMappingDecoder(upperCaseResponseMapper(), new StringDecoder());
+    String output = (String) decoder.decode(responseWithText("response"), String.class);
+
+    assertThat(output).isEqualTo("RESPONSE");
+  }
+
+  private ResponseMapper upperCaseResponseMapper() {
+    return new ResponseMapper() {
+      @Override
+      public Response map(Response response, Type type) {
+        try {
+          return response
+                  .toBuilder()
+                  .body(Util.toString(response.body().asReader()).toUpperCase().getBytes())
+                  .build();
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
+      }
+    };
+  }
+
+  private Response responseWithText(String text) {
+    return Response.builder()
+            .body(text, Util.UTF_8)
+            .status(200)
+            .headers(new HashMap<String, Collection<String>>())
+            .build();
+  }
+
+  @Test
+  public void mapAndDecodeExecutesMapFunction() {
+    server.enqueue(new MockResponse().setBody("response!"));
+
+    TestInterface api = new Feign.Builder()
+            .mapAndDecode(upperCaseResponseMapper(), new StringDecoder())
+            .target(TestInterface.class, "http://localhost:" + server.getPort());
+
+    assertEquals(api.post(), "RESPONSE!");
+  }
+
   interface TestInterface {
 
     @RequestLine("POST /")
@@ -715,6 +789,9 @@ public class FeignTest {
     @RequestLine("GET /")
     void queryMap(@QueryMap Map<String, Object> queryMap);
 
+    @RequestLine("GET /")
+    void queryMapEncoded(@QueryMap(encoded = true) Map<String, Object> queryMap);
+
     @RequestLine("GET /?name={name}")
     void queryMapWithQueryParams(@Param("name") String name, @QueryMap Map<String, Object> queryMap);
 
@@ -763,6 +840,17 @@ public class FeignTest {
     @Override
     public Exception decode(String methodKey, Response response) {
       if (response.status() == 400) {
+        return new IllegalArgumentException("bad zone name");
+      }
+      return super.decode(methodKey, response);
+    }
+  }
+
+  static class IllegalArgumentExceptionOn404 extends ErrorDecoder.Default {
+
+    @Override
+    public Exception decode(String methodKey, Response response) {
+      if (response.status() == 404) {
         return new IllegalArgumentException("bad zone name");
       }
       return super.decode(methodKey, response);